Refactor and adds some regression tests
